The previous version of ""simulatedyn" had a bug where it was not updating only one step but many so the produced networks were wrong. I have fixed it (the last time!).
"simulatedyn" alo has a new option called "diff" which if TRUE returns a
network series of sequential change networks in the network rather than the
sequential networks themselves. That is, the edgelists in the returned network
series are those dyads that changed during that step. The reason is that the
size of these is very small for small steps. The new function "network.toggle"
updates the network:

nw1 <- network.toggle(nw0, diffnetwork)

updates nw0 to nw1 by toggling the dyads that are edges in the network
diffnetwork. This is much faster than a regular update.

"simulatedyn" now uses much less RAM, disk storage and is faster.


git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/statnet/trunk@16 9cb281f4-553f-406f-8d7f-11f072bbfe46