I have fixed a number of minor bugs.
i have added a new summary function "Summary.ergm" that is closer to
"summary.lm" in that it uses a print method. But it is different from the
current function in the formatting of
its output (mainly the order of the mcmc se and p-values have been swapped).
So I have kept it separate for now.

I have also made "Summary.ergm" return a list of
summary numbers like the AIC and BIC.

I have made "simulatedyn" return a "network series" as a list with two
components:

$changed: a three column matrix (time step, head node, tail node)
  There is one row for each changed dyad over the entire series (i.e., time
0 to the last time). Values listed under time step i are the changes made to
the network at time step i to make it the network at time i+1. They are
listed in time step order.

$dissolved:  a three column matrix (time step, head node, tail node)
  There is one row for each dissolved dyad over the entire series (i.e.,
time 0 to the last time). Values listed under time step i are the ties
dissolved at time step i to make the network at time i+1. They are listed in
time step order. This differs from $changed as a tie may be dissolved and
reformed in the same time step. Such ties will be on in $dissolved but not
$changed. Of course, $changed also list the formed dyads.

The networks themselves can be reconstructed using code like:

gs <- g0
for(i in 1:numtimesteps){
  gs <- network.toggle(gs,gsim,time=i-1)
}

where gs is sequential the network at time steps 0, 1, ..., numtimesteps,
and gsim is the result of "simulatedyn". Note that the original
"network.toggle(nw, edgelist)" still works as it tests for a "time"
argument.

The dynamic code is now all in C, uses modest memory and storage and is
lightning fast (e.g., 15 seconds to generate 10000 time steps). The
reconstructions are fast also. This is a substantial step forward in
usability.


git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/statnet/trunk@18 9cb281f4-553f-406f-8d7f-11f072bbfe46