This is a brief desription of PRIOR changes to "statnet" since the 1.0-0 release before the Sunbelt. This was reported in an email on 7/21/2006 and here for the record. I can not recall a bug being reported except for one that I found and note below. However, there have been many requests for enhancements, some of the easy ones I have made.
This refer to version is 1.0-4

The changes:

1) I implemented the "boundDeg" term (to allow constraints on the in/out degree
of actors). It was documented but not implemented.

2) I implemented the "altkstar(lambda)" term for the "alternating k-star"
statistic of eq. (13), p. 11, of the "New Specifications" paper. It is
available for undirected networks only.

3) I reorganized the "geometrically weighted degree" terms.

a) "gwdegree(decay)" is exactly eq. (14) in Dave's Sunbelt paper
[http://www.sna.unimelb.edu.au/publications/cef4.pdf] (decay is \theta_s in eq.
(14)).

b) "gwdegreealpha(alpha)" was the original "gwdegree" in 1.0-0.

c)  gwdegreelambda(lambda)" is eq. (14) in Dave's Sunbelt paper (lambda is
exp(\theta_s) in eq. (14)). This parameterization make it close to the
"altkstar".

All versions allow estimation of the non-linear parameter.

This was because we (Dave and I) prefer his final parameterization, and want to
clarify things given the variants in various papers. Peter Marsden was trying
to use it and was (rightly) confused. I have documented "altkstar"
and "gwdegree" but not the others as they are legacy only.

4) I found a few documentation very minor bugs that I fixed.

6) I added a "simmelian" term at the request of David Krackhardt. More on that
later.

7) The bug I found occurs when two curved terms are included in the same model.
For example:

library(statnet)
data(florentine)

flobusiness.fit <- ergm(flobusiness ~ edges + gwesp(0.2) + gwdegree(0.5),
burnin=1000, MCMCsamplesize=5000, interval=1000, maxit=1)

Produces the error after the MCMC of:

Error in if (ec[istat] > 0) { : missing value where TRUE/FALSE needed

It seems the "canonical" parameter is not being computed correctly.

Dave: Can you help here? (I think he did)

Overall, these are minor changes.

It appears that the release has been successful and a number of key people are
adopting it. I think we should move slowly on developing it (given our other
responsibilities) and think about moving over other functionality from the
internal version (e.g., bipartite, valued edges, missing data) for the next
Sunbelt.


git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/statnet/trunk@64 9cb281f4-553f-406f-8d7f-11f072bbfe46