Dave H's changes Mar. 25 2007:
Added gwadegree and gwedegree terms.  Have not yet implemented the curved exponential family
capability with these functions, but they do allow for degree by attribute.

Fixed problem with calculation of change statistics in gwdegree, gwodegree, and gwidegree terms
(This was a serious problem but it was easy to change with minor mods to d_gwdegree, d_gwodegree,
and d_gwidegree.)



git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/statnet/trunk@65 9cb281f4-553f-406f-8d7f-11f072bbfe46