Added various terms to measure the R_0 of a network based on the "random mixing by degree" model. See Handcock and Jones 2006 for details.
The new term "r0" has no arguments and computes the R_0 of the network when the integrated
transmissibility is 1. The R_0 is proportional to the integrated transmissibility so it can be
computed for any other value of the integrated transmissibility.

As these terms are essentially a measure of dispersion of the mommentary
degree distribution, "r0" is another term to measure the effect of concurrency.

It works for both monopartite and bipartite (undirected) networks.

The terms "r0a" and "r0e" compute R_0 for the actor and event distributions of a bipartite
network. The R_0 for the network as a whole is (integrated transmissibility)*r0a*r0e. 


git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/statnet/trunk@68 9cb281f4-553f-406f-8d7f-11f072bbfe46