Added a simulated annealing algorithm. It is called just like "ergm"
It samples from theta0 toggles and makes them if they reduce the Euclidean distance of the network from the
"meanstats" parameter.

If theta0 is not given the  MPLE is used.

This is useful to get a network that has statistics essentially a given set.
If the "burnin" is large it is very close to meanstats.


git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/statnet/trunk@120 9cb281f4-553f-406f-8d7f-11f072bbfe46