This is an aggregation of many minor changes I have made over the last few months to my local version of ergm. The major changes are:
1) Robustified the computation of the MCMC s.e. so it does not bomb as much.
2) In ergm.estimate I added a few options for alternative approximations to the likelihood. These are specified via the metric argument to the control.ergm function:

metric=c("Likelihood","Median.Likelihood","EF.Likelihood")
"Likelihood" = The log-normal approximation to the likelihood using the mean and variance to approximate the Normal parameters.
"Median.Likelihood" = The log-normal approximation to the likelihood using the median and IQR to approximate the Normal parameters.
"EF.Likelihood" = The direct computation ofthe likelihood using the original Geyer-Thompson form.

The new default is "Likelihood" while the old one was "Median.Likelihood". The "Likelihood" appears to perform better and be more stable.

I also made the computation of the gradients better and used the gradient functions in the optim() call. This seems to help convergence also.

I also reduced the convergence level of optim() so it thinks it has converged sooner. As we have a rough likelihood it should not be looking for convergence to 10^-8.

I can change these back if it bothers folks or if these "tweaks" do not work in cases we  care about..



git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/trunk@10774 9cb281f4-553f-406f-8d7f-11f072bbfe46