Added a new option to "steplength" in control.ergm called 'steplength="adaptive"' (The default remains 0.5) which starts with a steplength of 1 and checks to see if the change in the log-likelihood is less than 2. If it is it finishes the iteration. If not it divides the steplength by 2 and repeats the change in the log-likelihood test until it passes (as it must).
Secondly, if steplength=1 and the change in the log-likelihood is less than 0.001 it does not do any more iterations (i.e., stops early).

The objective is  to make it more adaptive and quicker. I am sure that Ruth and Dave have a better criterion, but this is useful for serious applications I am doing now.



git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/trunk@10781 9cb281f4-553f-406f-8d7f-11f072bbfe46