Added major new MPLE option for the case where "constraints=~degree"
This is different from the usual dyadic MPLE as it requires a PL with constant degree sequence (which the MPLE explicitly breaks). The basic idea is to consider the set of tetrads and hexads consistent with the  degree sequence and construct a "composite" likelihood from them. The technical and computational aspects are too detailed to be scribbled into the margins of this log. Incredibly, it seems to work very well.
To use it just set "MPLEonly=TRUE" when using "constraints=~degree". That is, I have integrated it into the regular MPLE/ergm framework.
This is worth adding to the CRAN version as it is a very nice new capability
that is also important for the dymanic models that condition on the
degree sequence.



git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/trunk@10784 9cb281f4-553f-406f-8d7f-11f072bbfe46