Changed the code back so compression is used all the time.
I believe there is a bug in the compressed code that effects the case where there is an offset and in the uncompressed code otherwise but I can not find it.

So I have used the compressed code and will try to find what is wrong with the offset code (if anything).

git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/trunk@10841 9cb281f4-553f-406f-8d7f-11f072bbfe46