Based on Pavel's identification of a problem with convergence for tiny networks, I have changed the default estimation method to "lognormal" from "Median.likelihood"
The "Median.likelihood" is very poor is the statsmatrix is very discrete as it approximates the mean of the lognormal by the (weighted) median of the sample statistics.
This is robust but subject to bias if the distribution is over very few point values (as in Pavel's tiny network - the number of edges was 0, 1 or 2 only).

Also changed the default covariance matrix method to the regular one (Pearson) from the robust one. The robust one is quite poor for discrete statsmatrices.

This makes the defaults less robust to extremely non-normal samples but more robust to tiny networks. Based on my experience this is a better compromise.

git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/trunk@10940 9cb281f4-553f-406f-8d7f-11f072bbfe46