Per enhancement ticket   https://statnet.csde.washington.edu/trac/ticket/512
Add ergm-constraint term fixedas(present,absent). 

fixeds preserves the edges in 'present' and precludes the edges in 'absent'. 
      Both 'present' and 'absent' can take input object as edgelist or network, the latter will convert to the corresponding edgelist.
      
The associated MHp method uses randomtoggle_list, hence it is not a TNT type proposal. I am working on the TNT type, however it is a bit tricky. 

 Add helpfile in ergm-constraints and add test file in constraint_generalized.R
 
fixallbut term (explained in the ticket) should be ready soon. 

git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/trunk@13348 9cb281f4-553f-406f-8d7f-11f072bbfe46