Pavel modified .Hummel.steplength() when there is missing data. he did this by checking if each point of the constrained sample was in the convex hull of the unconstrained sample. As the samples are large lp() from lpSolve was called for each constrained sample point, and this takes a long time.
I modified the code to approximate the convex hull of the  constrained sample points by finding the 100 points with the lowest probability under the hypothetical Gaussian model for them. These should be the extreme points. These 100 points are checked in order of extremeness. They will be inside the true hull but this should be close enough.

This is only implemented if the number of search points is greater than 100 (i.e., only in the missing data case).

The old method used a uniroot() to find the limit. I changed this to do a Bayesian search with the prior on the step length (truncated) gaussian centered on the previous value and with s.d. 0.1. Each test of a step length leads to a posterior that truncates the distribution (i.e. resets the lower or upper bound). For each step I choose the posterior median as the next test point and repeated the search. I set a tolerance of 0.01 or at most 20 iterations. This seems to work well and I have made it the default for all cases.
 

git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/branches/NewHummelSearch@14253 9cb281f4-553f-406f-8d7f-11f072bbfe46