Robustify .find_OK_burnin in rare cases of pathological sampling statistics.
This is a recommit of a reverted commit sent in error.

It is to catch cases where the sampling statistics are near constant so that the optimization fails. These changes enable it to recover and increase the burn-in for a retry rather than fail.
It occurred during a complicated tapered model fit.