This is a transference of the ergm-private" "control_llik" branch to the public version ergm 4.0.
The branch consolidates control arguments to the likelihood functions used in ergm.estimate into a single list variable called "control.llik".
This simplifies the calls to the functions.

The branch also adds a simple hook in "ergm.estimate" for the Kpenalty likelihood functions. These are required for the "ergm'tapered" package to interoperate with "ergm".

Also fixed a few documentation typos in "ergm".