Added a MCMC.esteq.exclude.statistics control argument
control$MCMC.esteq.exclude.statistics allows network statistics to not be included in the estimating equations. There equations determine most convergence diagnostics. For some models there are some terms that are not zero at convergence and should be removed.  For example, this supports tapering models which may have terms with non-specific means. The estimating equations are used in a number of places and the code follows those places.

Specifically, it is a vector of  strings being the names of the statistics to be excluded from
the estimating equations.